# CadastroFilmes/Models/Genero.cs
using System;


namespace CadastroFilmes.Models
{
    public class Genero
    { 
        public int Id { get; set; }
        public string Nome { get; set; }
        public DateTime DataCriacao { get; set; }
        public int Ativo { get; set; }
     
    }
}