# CadastroFilmes/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CadastroFilmes",
    "version": "v1"
  },
  "paths": {
    "/api/Filmes": {
      "get": {
        "tags": [
          "Filme"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Filme"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Filmes/{id}": {
      "get": {
        "tags": [
          "Filme"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Filme"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Filme"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Filmes/{ids}": {
      "delete": {
        "tags": [
          "Filme"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "type": "array",
            "items": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "ids",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Genero": {
      "get": {
        "tags": [
          "Genero"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Genero"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Genero"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Genero/{id}": {
      "get": {
        "tags": [
          "Genero"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Genero": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "dataCriacao": {
          "format": "date-time",
          "type": "string"
        },
        "ativo": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Filme": {
      "type": "object",
      "properties": {
        "filmeId": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "dataCriacao": {
          "format": "date-time",
          "type": "string"
        },
        "ativo": {
          "format": "int32",
          "type": "integer"
        },
        "generoId": {
          "format": "int32",
          "type": "integer"
        },
        "genero": {
          "$ref": "#/definitions/Genero"
        }
      }
    }
  }
}